feat: divide to small Input components
